Fix multilevel artifact downloads
Models can be registered with subfolders, for example such as:
```
test
├── subdir
│   └── sub.file
└── top.file
```

In the current form, when one requests this model, the library downloads
all files, but the files from the subdirectories are downloaded
multiple times: in the root, and also in their subfolder (and along all
the subfolders, if there are even more levels):
```
test
├── subdir
│   └── sub.file
├── sub.file
└── top.file
```
This is because `mlflow` expects a list of files & directories in a path,
but doesn't expect all the items recursively within a subdir, since there
are extra steps when they traverse those subdirectories.

With this change, when `mlflow` requests a list of artifacts at a given
path, only return the files and directories directly there, and filter
out all other recursive items (which `mlflow` will request later on).

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>